Moved implementation of kernels out to dedicated header files.
