+++ dpctl/tensor/libtensor/include/utils/type_utils.hpp
@@ -1,0 +1,73 @@
+//===------ type_utils.hpp - Implementation of types utils  ----*-C++-*/===//
+//
+//                      Data Parallel Control (dpctl)
+//
+// Copyright 2020-2022 Intel Corporation
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+//===----------------------------------------------------------------------===//
+///
+/// \file
+/// This file defines functions for value casting.
+//===----------------------------------------------------------------------===//
+
+#pragma once
+#include <complex>
+
+namespace dpctl
+{
+namespace tensor
+{
+namespace type_utils
+{
+
+template <class T> struct is_complex : std::false_type
+{
+};
+template <class T> struct is_complex<std::complex<T>> : std::true_type
+{
+};
+
+template <typename dstTy, typename srcTy> dstTy convert_impl(const srcTy &v)
+{
+    if constexpr (std::is_same<dstTy, srcTy>::value) {
+        return v;
+    }
+    else if constexpr (std::is_same_v<dstTy, bool> && is_complex<srcTy>::value)
+    {
+        // bool(complex_v) == (complex_v.real() != 0) && (complex_v.imag() !=0)
+        return (convert_impl<bool, typename srcTy::value_type>(v.real()) ||
+                convert_impl<bool, typename srcTy::value_type>(v.imag()));
+    }
+    else if constexpr (is_complex<srcTy>::value && !is_complex<dstTy>::value) {
+        // real_t(complex_v) == real_t(complex_v.real())
+        return convert_impl<dstTy, typename srcTy::value_type>(v.real());
+    }
+    else if constexpr (!std::is_integral<srcTy>::value &&
+                       !std::is_same<dstTy, bool>::value &&
+                       std::is_integral<dstTy>::value &&
+                       std::is_unsigned<dstTy>::value)
+    {
+        // first cast to signed variant, the cast to unsigned one
+        using signedT = typename std::make_signed<dstTy>::type;
+        return static_cast<dstTy>(convert_impl<signedT, srcTy>(v));
+    }
+    else {
+        return static_cast<dstTy>(v);
+    }
+}
+
+} // namespace type_utils
+} // namespace tensor
+} // namespace dpctl
